+++ klippy/extras/extruder_stepper.py
@@ -4,36 +4,17 @@
 #
 # This file may be distributed under the terms of the GNU GPLv3 license.
 import logging
-import stepper
+from kinematics import extruder
 
-class ExtruderStepper:
+class PrinterExtruderStepper:
     def __init__(self, config):
         self.printer = config.get_printer()
-        stepper_name = config.get_name().split()[1]
+        self.extruder_stepper = extruder.ExtruderStepper(config)
         self.extruder_name = config.get('extruder', 'extruder')
-        self.stepper = stepper.PrinterStepper(config)
-        self.stepper.setup_itersolve('extruder_stepper_alloc')
         self.printer.register_event_handler("klippy:connect",
                                             self.handle_connect)
-        gcode = self.printer.lookup_object('gcode')
-        gcode.register_mux_command("SYNC_STEPPER_TO_EXTRUDER", "STEPPER",
-                                   stepper_name,
-                                   self.cmd_SYNC_STEPPER_TO_EXTRUDER,
-                                   desc=self.cmd_SYNC_STEPPER_TO_EXTRUDER_help)
     def handle_connect(self):
-        extruder = self.printer.lookup_object(self.extruder_name)
-        extruder.sync_stepper(self.stepper)
-        toolhead = self.printer.lookup_object('toolhead')
-        toolhead.register_step_generator(self.stepper.generate_steps)
-    cmd_SYNC_STEPPER_TO_EXTRUDER_help = "Set extruder stepper"
-    def cmd_SYNC_STEPPER_TO_EXTRUDER(self, gcmd):
-        ename = gcmd.get('EXTRUDER')
-        extruder = self.printer.lookup_object(ename, None)
-        if extruder is None:
-            raise gcmd.error("'%s' is not a valid extruder." % (ename,))
-        extruder.sync_stepper(self.stepper)
-        self.extruder_name = ename
-        gcmd.respond_info("Extruder stepper now syncing with '%s'" % (ename,))
+        self.extruder_stepper.sync_to_extruder(self.extruder_name)
 
 def load_config_prefix(config):
-    return ExtruderStepper(config)
+    return PrinterExtruderStepper(config)
